# docs/swagger.yaml
basePath: /
definitions:
  api.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  api.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  api.SubscriptionResponse:
    properties:
      created_at:
        type: string
      currency:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  api.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления подписками
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает список подписок с пагинацией и фильтрацией
      parameters:
      - description: Фильтр по user_id
        in: query
        name: user_id
        type: string
      - description: Фильтр по service_name
        in: query
        name: service_name
        type: string
      - default: 10
        description: Лимит (по умолчанию 10)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SubscriptionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку для пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/api.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Получить подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет существующую подписку
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/api.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Возвращает общую стоимость подписок за период
      parameters:
      - description: Фильтр по user_id
        in: query
        name: user_id
        type: string
      - description: Фильтр по service_name
        in: query
        name: service_name
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: period_start
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: period_end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              from:
                type: string
              service:
                type: string
              timestamp:
                type: string
              to:
                type: string
              total:
                format: int64
                type: integer
              user_id:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIResponse'
      summary: Сумма подписок
      tags:
      - subscriptions
swagger: "2.0"