# pkg/api/subscription.go
package api

// SubscriptionResponse represents subscription data in API response
// swagger:model SubscriptionResponse
type SubscriptionResponse struct {
	ID          string `json:"id"`
	ServiceName string `json:"service_name"`
	Price       int64  `json:"price"`
	Currency    string `json:"currency"`
	UserID      string `json:"user_id"`
	StartDate   string `json:"start_date"`
	EndDate     string `json:"end_date"`
	CreatedAt   string `json:"created_at"`
	UpdatedAt   string `json:"updated_at"`
}